# services/data-service/src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Escape Doom Data API
  version: 0.0.1
servers:
  - url: https://api.server.test/v1/
paths:
...
  /scene:
    get:
      operationId: getAllScenes
      tags:
        - Scene
      summary: Get all scenes
      description: Retrieve all scenes that are not linked to any specific level
      responses:
        '200':
          description: A list of scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SceneDTO'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createScene
      tags:
        - Scene
      summary: Create a new scene
      description: Create a Scene independently of any level
      requestBody:
        description: The details of the new Scene
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneRequestDTO'
      responses:
        '201':
          description: Scene created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /scenes/{escape-room-scene-id}:
    get:
      operationId: getScene
      tags:
        - Scene
      summary: Get details of a scene
      description: Retrieve details of a specific scene by its ID
      parameters:
        - name: escape-room-scene-id
          in: path
          required: true
          description: The unique ID of the Scene
          schema:
            type: string
            example: scene123
      responses:
        '200':
          description: Scene details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneDTO'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      operationId: putScene
      tags:
        - Scene
      summary: Update a scene
      description: Update the details of a specific Scene
      parameters:
        - name: escape-room-scene-id
          in: path
          required: true
          description: The unique ID of the Scene
          schema:
            type: string
            example: scene123
      requestBody:
        description: The updated details of the Scene
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SceneRequestDTO'
      responses:
        '200':
          description: Scene updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SceneDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      operationId: deleteScene
      tags:
        - Scene
      summary: Delete a scene
      description: Delete a specific Scene by its ID
      parameters:
        - name: escape-room-scene-id
          in: path
          required: true
          description: The unique ID of the Scene
          schema:
            type: string
            example: scene123
      responses:
        '200':
          description: Scene deleted successfully
          content:
            application/json:
              schema:
                type: object
                title: DeleteLevelResponseDTO
                properties:
                  message:
                    type: string
                    example: Scene deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
    NodeType:
      description: The type of a node
      type: string
      enum:
        - ZOOM
        - CONSOLE
        - DETAIL
        - STORY
      example: "ZOOM"
    NodeDTO:
      description: A node of an escape-room instance
      type: object
      properties:
        node_type:
          $ref: '#/components/schemas/NodeType'
        node_info:
          $ref: '#/components/schemas/NodeInfoDTO'
        position:
          $ref: '#/components/schemas/PositionDTO'
    NodeInfoDTO:
      description: The information of a node
      type: object
      properties:
        description:
          type: string
          description: The description of the node
          example: This is a story node
        title:
          type: string
          description: The display title of the node
          example: I like cheese
        imageURI:
          type: string
          description: The URI of the image to display
          example: https://example.com/image.png
    PositionDTO:
      description: The position of a node
      type: object
      properties:
        top:
          type: number
          format: double
          description: The relative top position of the node
          example: 50.5
        left:
          type: number
          format: double
          description: The relative left position of the node
          example: 22


    # Enums
...
    SceneDTO:
      description: A scene of an escape-room instance
      type: object
      properties:
        escape_room_level_id:
          type: string
          description: The ID of the escape room level that contains this riddle
          example: "a12b34c5-6789-4def-abcd-12345678abcd"
        escape_room_sequence_id:
          type: string
          description: The unique ID of the scene sequence
          example: "241a70fe-47d6-4756-9ac7-330f1b199e84"
        scene_sequence:
          type: integer
          description: Defines the "position" of the scene in the escape room
          example: 1
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeDTO"
          description: The nodes of the scene
        background_image_uri:
          type: string
          description: The URI of the background image
          example: "https://example.com/background.png"
        name:
          type: string
          description: The name of the scene
          example: "Scene 1"
...
    SceneRequestDTO:
      description: Base schema for any request to the scene API
      type: object
      properties:
        scene_sequence:
          type: integer
          description: Defines the "position" of the scene in the escape room
          example: 1
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeDTO"
          description: The nodes of the scene
        background_image_uri:
          type: string
          description: The URI of the background image
          example: "https://example.com/background.png"
        name:
          type: string
          description: The name of the scene
          example: "Scene 1"
        escape_room_level_id:
          type: string
          description: The ID of the Escape Room Level this scene belongs to
          example: "c2d1a3b4-5e6f-47b8-9c9d-0a1b2c3d4e5f"
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            title: CreateBadRequestDTO
            properties:
              message:
                type: string
                example: Invalid data provided
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            title: CreateNotFoundDTO
            properties:
              message:
                type: string
                example: The requested resource was not found
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            type: object
            title: CreateInternalServerErrorDTO
            properties:
              message:
                type: string
                example: An unexpected error occurred on the server
...